Move the cart actions to the store branch.
