+++ web/app/containers/mini-cart/actions.js
@@ -1,4 +1,4 @@
-import {getCart} from '../cart/actions'
+import {getCart} from '../../store/cart/actions'
 import * as modalActions from '../../store/modals/actions'
 import {MINI_CART_MODAL} from './constants'
