+++ web/app/containers/mini-cart/reducer.js
@@ -1,11 +1,11 @@
 import {handleActions} from 'redux-actions'
 import Immutable from 'immutable'
-import * as cartActions from '../cart/actions'
+import {receiveCartContents} from '../../store/cart/actions'
 
 const initialState = Immutable.fromJS({
     contentsLoaded: false,
 })
 
 export default handleActions({
-    [cartActions.receiveCartContents]: (state) => state.set('contentsLoaded', true)
+    [receiveCartContents]: (state) => state.set('contentsLoaded', true)
 }, initialState)
